Rename CustomSerde package and remove unused validatedCustomSerde logic in RuleLambdaGenerator.
